+++ venus-shared/typeutil/fields_test.go
@@ -1,4 +1,4 @@
-package typecheck
+package typeutil
 
 import (
 	"testing"
