+++ venus-shared/typeutil/methods_test.go
@@ -1,4 +1,4 @@
-package typecheck
+package typeutil
 
 import (
 	"io"
