+++ venus-shared/typeutil/similar.go
@@ -1,4 +1,4 @@
-package typecheck
+package typeutil
 
 import (
 	"encoding"
